break out update logic into its own service to call from the release controller, markup with otel activities
